Delete Config.ts and update start script in package.json
