+++ InfrastructureAgent/Index.ts
@@ -1,99 +1,25 @@
-import BasicCron from 'CommonServer/Utils/BasicCron';
-import { BasicMetircs } from './Utils/BasicMetrics';
-import { EVERY_MINUTE } from 'Common/Utils/CronTime';
-import logger from 'CommonServer/Utils/Logger';
-import API from 'Common/Utils/API';
+import yargs from 'yargs';
 import URL from 'Common/Types/API/URL';
-import BaseModel from 'Common/Models/BaseModel';
-import Monitor from 'Model/Models/Monitor';
-import ServerMonitorResponse from 'Common/Types/Monitor/ServerMonitor/ServerMonitorResponse';
-import OneUptimeDate from 'Common/Types/Date';
-import JSONFunctions from 'Common/Types/JSONFunctions';
-import HTTPErrorResponse from 'Common/Types/API/HTTPErrorResponse';
-import HTTPResponse from 'Common/Types/API/HTTPResponse';
-import { OneUptimeURL, SecretKey } from './Utils/Config';
-import BadDataException from 'Common/Types/Exception/BadDataException';
+import Dictionary from 'Common/Types/Dictionary';
+import MonitorInfrastructure from './Jobs/MonitorInfrastructure';
 
-BasicCron({
-    jobName: 'MonitorInfrastructure',
-    options: {
-        schedule: EVERY_MINUTE, // Every minute
-        runOnStartup: true,
-    },
-    runFunction: async () => {
-        try {
-            const secretKey: string | undefined = SecretKey;
-            const oneuptimeHost: URL = OneUptimeURL;
+const usage = "\nUsage: oneuptime-infrastructure-agent --secret-key <secret-key>";
 
-            if (!secretKey) {
-                throw new BadDataException(
-                    'No SECRET_KEY environment variable found. You can find secret key for this monitor on OneUptime Dashboard'
-                );
-            }
+const argv: Dictionary<string> = yargs  
+      .usage(usage)  
+      .option("k", {alias:"secret-key", describe: "Secret Key for this agent. You will find this on OneUptime Dashboard", type: "string", demandOption
+: true })
+    .option("h", {alias:"oneuptime-host", describe: "OneUptime Host. By default this is https://oneupime.com", type: "string", demandOption
+    : false })                                                                                                     
+      .help(true)  
+      .argv as Dictionary<string>;
 
-            // get monitor steps to get disk paths.
-            const monitorResult: HTTPErrorResponse | HTTPResponse<BaseModel> =
-                await API.get(
-                    URL.fromString(
-                        `${oneuptimeHost}/server-monitor/${secretKey}`
-                    )
-                );
+const secretKey: string | undefined = argv["secret-key"];
+const oneuptimeHost: URL = URL.fromString(argv["oneuptime-host"] || "https://oneuptime.com");
 
-            if (monitorResult instanceof HTTPErrorResponse) {
-                throw monitorResult;
-            }
 
-            const monitor: Monitor = BaseModel.fromJSON(
-                monitorResult.data,
-                Monitor
-            ) as Monitor;
+if(!secretKey) {
+    throw new Error("No secret-key argument found. You can find secret key for this monitor on OneUptime Dashboard");
+}
 
-            // get disk paths to monitor.
-
-            const diskPaths: string[] = [];
-
-            for (const step of monitor.monitorSteps?.data
-                ?.monitorStepsInstanceArray || []) {
-                for (const criteriaInstance of step.data?.monitorCriteria.data
-                    ?.monitorCriteriaInstanceArray || []) {
-                    for (const filter of criteriaInstance.data?.filters || []) {
-                        if (filter.serverMonitorOptions?.diskPath) {
-                            diskPaths.push(
-                                filter.serverMonitorOptions?.diskPath
-                            );
-                        }
-                    }
-                }
-            }
-
-            const serverMonitorResponse: ServerMonitorResponse = {
-                monitorId: monitor.id!,
-                requestReceivedAt: OneUptimeDate.getCurrentDate(),
-                basicInfrastructureMetrics: await BasicMetircs.getBasicMetrics({
-                    diskPaths: diskPaths,
-                }),
-                onlyCheckRequestReceivedAt: false,
-            };
-
-            logger.info('Server Monitor Response');
-            logger.info(serverMonitorResponse);
-
-            // now we send this data back to server.
-
-            await API.post(
-                URL.fromString(
-                    `${oneuptimeHost}/server-monitor/response/ingest/${secretKey}`
-                ),
-                {
-                    serverMonitorResponse: JSONFunctions.serialize(
-                        serverMonitorResponse as any
-                    ),
-                },
-                {}
-            );
-        } catch (err) {
-            logger.error('Error reporting metrics to OneUptime Server.');
-            logger.error(err);
-        }
-    },
-});
+MonitorInfrastructure.initJob(secretKey, oneuptimeHost);
